move server code to separate directory
